pkg/k8s: move pkg/k8s/cep.go to pkg/k8s/endpointsynchronizer/cep.go
This avoids pkg/k8s to import pkg/endpoint which is used by
cilium/operator.
