+++ pkg/k8s/endpointsynchronizer/logfields.go
@@ -1,0 +1,31 @@
+// Copyright 2016-2018 Authors of Cilium
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package endpointsynchronizer
+
+import (
+	"github.com/cilium/cilium/pkg/logging"
+	"github.com/cilium/cilium/pkg/logging/logfields"
+)
+
+// logging field definitions
+const (
+	// subsysEndpointSync is the value for logfields.LogSubsys
+	subsysEndpointSync = "endpointsynchronizer"
+)
+
+var (
+	// log is the endpointsynchronizer package logger object.
+	log = logging.DefaultLogger.WithField(logfields.LogSubsys, subsysEndpointSync)
+)
